backend/requirements.txt
flask==3.1.*
bcrypt==4.2.*
PyJWT==2.9.*
flask-cors==5.0.*
python-dotenv==1.1.*
